+++ sample/proto_v3/role_cfg.amd.js
@@ -2,9 +2,9 @@
 
 define({
   role_cfg_header: {
-    xres_ver : "1.4.1",
+    xres_ver : "1.4.2",
     hash_code : "no hash code",
-    data_ver : "1.4.1.20171206065119",
+    data_ver : "1.4.2.20180317091902",
     count : 7
   },
   role_cfg: [